custom sorting hook isolated in its own module
